multi-llm-router/requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: For FastAPI integration
# fastapi>=0.100.0
# uvicorn>=0.22.0